+++ core/database/src/main/java/com/hellguy39/hellnotes/database/entity/RemindEntity.kt
@@ -9,8 +9,8 @@ const val REMINDS_TABLE_NAME = "reminds_table"
 
 @Entity(tableName = REMINDS_TABLE_NAME)
 data class RemindEntity(
-    @PrimaryKey(autoGenerate = true) val id: Int? = null,
-    val noteId: Int,
+    @PrimaryKey(autoGenerate = true) val id: Long? = null,
+    val noteId: Long,
     val message: String,
     val triggerDate: Long
 )
