+++ core/model/src/main/java/com/hellguy39/hellnotes/model/Remind.kt
@@ -1,8 +1,8 @@
 package com.hellguy39.hellnotes.model
 
 data class Remind(
-    val id: Int? = null,
-    val noteId: Int,
+    val id: Long? = null,
+    val noteId: Long,
     val message: String,
     val triggerDate: Long
 )
